adjust trips/trip/:id to show driver and customer name, change payment to numeric

+++ backend/backend-express/db/schema/01_schema.sql
@@ -43,7 +43,7 @@ CREATE TABLE trips(
   end_location_lat REAL NOT NULL,
   end_location_lon REAL NOT NULL,
   accepted boolean DEFAULT FALSE,
-  payment_amount INTEGER DEFAULT 0,
+  payment_amount NUMERIC(6, 2) DEFAULT 0,
   payment_status boolean NOT NULL DEFAULT FALSE,
   created_at timestamp DEFAULT NOW(),
   ended_at timestamp DEFAULT NULL
